Product inplemented

+++ src/Authorization.Desktop/ViewModels/Products/ProductAllToStorageViewModel.cs
@@ -8,6 +8,6 @@ public class ProductAllToStorageViewModel
     public string Name { get; set; } = string.Empty;
     public long BarCode { get; set; }
     public long Quantity { get; set; }
-    public float SoldPrice { get; set; }
-    public float Price { get; set; }
+    public long SoldPrice { get; set; }
+    public long Price { get; set; }
 }
