+++ src/Authorization.Desktop/ViewModels/Products/ProductViewModel.cs
@@ -9,8 +9,8 @@ public class ProductViewModel
     public string Name { get; set; } = string.Empty;
     public long BarCode { get; set; }
     public long Quantity { get; set; }
-    public float SoldPrice { get; set; }
-    public float Price { get; set; }
+    public long SoldPrice { get; set; }
+    public long Price { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set;}
 }
